execdetails: move ScanDetail and TimeDetail into store/tikv
