+++ sessionctx/stmtctx/stmtctx.go
@@ -542,19 +542,19 @@ func (sc *StatementContext) MergeExecDetails(details *execdetails.ExecDetails, c
 }
 
 // MergeScanDetail merges scan details into self.
-func (sc *StatementContext) MergeScanDetail(scanDetail *execdetails.ScanDetail) {
+func (sc *StatementContext) MergeScanDetail(scanDetail *util.ScanDetail) {
 	// Currently TiFlash cop task does not fill scanDetail, so need to skip it if scanDetail is nil
 	if scanDetail == nil {
 		return
 	}
 	if sc.mu.execDetails.ScanDetail == nil {
-		sc.mu.execDetails.ScanDetail = &execdetails.ScanDetail{}
+		sc.mu.execDetails.ScanDetail = &util.ScanDetail{}
 	}
 	sc.mu.execDetails.ScanDetail.Merge(scanDetail)
 }
 
 // MergeTimeDetail merges time details into self.
-func (sc *StatementContext) MergeTimeDetail(timeDetail execdetails.TimeDetail) {
+func (sc *StatementContext) MergeTimeDetail(timeDetail util.TimeDetail) {
 	sc.mu.execDetails.TimeDetail.ProcessTime += timeDetail.ProcessTime
 	sc.mu.execDetails.TimeDetail.WaitTime += timeDetail.WaitTime
 }
