+++ sessionctx/stmtctx/stmtctx_test.go
@@ -20,6 +20,7 @@ import (
 
 	. "github.com/pingcap/check"
 	"github.com/pingcap/tidb/sessionctx/stmtctx"
+	"github.com/pingcap/tidb/store/tikv/util"
 	"github.com/pingcap/tidb/util/execdetails"
 )
 
@@ -39,7 +40,7 @@ func (s *stmtctxSuit) TestCopTasksDetails(c *C) {
 			CalleeAddress: fmt.Sprintf("%v", i+1),
 			BackoffSleep:  make(map[string]time.Duration),
 			BackoffTimes:  make(map[string]int),
-			TimeDetail: execdetails.TimeDetail{
+			TimeDetail: util.TimeDetail{
 				ProcessTime: time.Second * time.Duration(i+1),
 				WaitTime:    time.Millisecond * time.Duration(i+1),
 			},
