+++ util/execdetails/execdetails.go
@@ -24,9 +24,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/pingcap/kvproto/pkg/kvrpcpb"
 	"github.com/pingcap/tidb/store/tikv/util"
-	"github.com/pingcap/tidb/util/memory"
 	"github.com/pingcap/tipb/go-tipb"
 	"go.uber.org/zap"
 )
@@ -42,8 +40,8 @@ type ExecDetails struct {
 	RequestCount     int
 	CommitDetail     *util.CommitDetails
 	LockKeysDetail   *util.LockKeysDetails
-	ScanDetail       *ScanDetail
-	TimeDetail       TimeDetail
+	ScanDetail       *util.ScanDetail
+	TimeDetail       util.TimeDetail
 }
 
 type stmtExecDetailKeyType struct{}
@@ -54,122 +52,6 @@ var StmtExecDetailKey = stmtExecDetailKeyType{}
 // StmtExecDetails contains stmt level execution detail info.
 type StmtExecDetails struct {
 	WriteSQLRespDuration time.Duration
-}
-
-// TimeDetail contains coprocessor time detail information.
-type TimeDetail struct {
-	// WaitWallTimeMs is the off-cpu wall time which is elapsed in TiKV side. Usually this includes queue waiting time and
-	// other kind of waitings in series.
-	ProcessTime time.Duration
-	// Off-cpu and on-cpu wall time elapsed to actually process the request payload. It does not
-	// include `wait_wall_time`.
-	// This field is very close to the CPU time in most cases. Some wait time spend in RocksDB
-	// cannot be excluded for now, like Mutex wait time, which is included in this field, so that
-	// this field is called wall time instead of CPU time.
-	WaitTime time.Duration
-}
-
-// String implements the fmt.Stringer interface.
-func (td *TimeDetail) String() string {
-	if td == nil {
-		return ""
-	}
-	buf := bytes.NewBuffer(make([]byte, 0, 16))
-	if td.ProcessTime > 0 {
-		buf.WriteString("total_process_time: ")
-		buf.WriteString(FormatDuration(td.ProcessTime))
-	}
-	if td.WaitTime > 0 {
-		if buf.Len() > 0 {
-			buf.WriteString(", ")
-		}
-		buf.WriteString("total_wait_time: ")
-		buf.WriteString(FormatDuration(td.WaitTime))
-	}
-	return buf.String()
-}
-
-// MergeFromTimeDetail merges time detail from pb into itself.
-func (td *TimeDetail) MergeFromTimeDetail(timeDetail *kvrpcpb.TimeDetail) {
-	if timeDetail != nil {
-		td.WaitTime += time.Duration(timeDetail.WaitWallTimeMs) * time.Millisecond
-		td.ProcessTime += time.Duration(timeDetail.ProcessWallTimeMs) * time.Millisecond
-	}
-}
-
-// ScanDetail contains coprocessor scan detail information.
-type ScanDetail struct {
-	// TotalKeys is the approximate number of MVCC keys meet during scanning. It includes
-	// deleted versions, but does not include RocksDB tombstone keys.
-	TotalKeys int64
-	// ProcessedKeys is the number of user keys scanned from the storage.
-	// It does not include deleted version or RocksDB tombstone keys.
-	// For Coprocessor requests, it includes keys that has been filtered out by Selection.
-	ProcessedKeys int64
-	// RocksdbDeleteSkippedCount is the total number of deletes and single deletes skipped over during
-	// iteration, i.e. how many RocksDB tombstones are skipped.
-	RocksdbDeleteSkippedCount uint64
-	// RocksdbKeySkippedCount it the total number of internal keys skipped over during iteration.
-	RocksdbKeySkippedCount uint64
-	// RocksdbBlockCacheHitCount is the total number of RocksDB block cache hits.
-	RocksdbBlockCacheHitCount uint64
-	// RocksdbBlockReadCount is the total number of block reads (with IO).
-	RocksdbBlockReadCount uint64
-	// RocksdbBlockReadByte is the total number of bytes from block reads.
-	RocksdbBlockReadByte uint64
-}
-
-// Merge merges scan detail execution details into self.
-func (sd *ScanDetail) Merge(scanDetail *ScanDetail) {
-	atomic.AddInt64(&sd.TotalKeys, scanDetail.TotalKeys)
-	atomic.AddInt64(&sd.ProcessedKeys, scanDetail.ProcessedKeys)
-	atomic.AddUint64(&sd.RocksdbDeleteSkippedCount, scanDetail.RocksdbDeleteSkippedCount)
-	atomic.AddUint64(&sd.RocksdbKeySkippedCount, scanDetail.RocksdbKeySkippedCount)
-	atomic.AddUint64(&sd.RocksdbBlockCacheHitCount, scanDetail.RocksdbBlockCacheHitCount)
-	atomic.AddUint64(&sd.RocksdbBlockReadCount, scanDetail.RocksdbBlockReadCount)
-	atomic.AddUint64(&sd.RocksdbBlockReadByte, scanDetail.RocksdbBlockReadByte)
-}
-
-var zeroScanDetail = ScanDetail{}
-
-// String implements the fmt.Stringer interface.
-func (sd *ScanDetail) String() string {
-	if sd == nil || *sd == zeroScanDetail {
-		return ""
-	}
-	buf := bytes.NewBuffer(make([]byte, 0, 16))
-	buf.WriteString("scan_detail: {")
-	buf.WriteString("total_process_keys: ")
-	buf.WriteString(strconv.FormatInt(sd.ProcessedKeys, 10))
-	buf.WriteString(", total_keys: ")
-	buf.WriteString(strconv.FormatInt(sd.TotalKeys, 10))
-	buf.WriteString(", rocksdb: {")
-	buf.WriteString("delete_skipped_count: ")
-	buf.WriteString(strconv.FormatUint(sd.RocksdbDeleteSkippedCount, 10))
-	buf.WriteString(", key_skipped_count: ")
-	buf.WriteString(strconv.FormatUint(sd.RocksdbKeySkippedCount, 10))
-	buf.WriteString(", block: {")
-	buf.WriteString("cache_hit_count: ")
-	buf.WriteString(strconv.FormatUint(sd.RocksdbBlockCacheHitCount, 10))
-	buf.WriteString(", read_count: ")
-	buf.WriteString(strconv.FormatUint(sd.RocksdbBlockReadCount, 10))
-	buf.WriteString(", read_byte: ")
-	buf.WriteString(memory.FormatBytes(int64(sd.RocksdbBlockReadByte)))
-	buf.WriteString("}}}")
-	return buf.String()
-}
-
-// MergeFromScanDetailV2 merges scan detail from pb into itself.
-func (sd *ScanDetail) MergeFromScanDetailV2(scanDetail *kvrpcpb.ScanDetailV2) {
-	if scanDetail != nil {
-		sd.TotalKeys += int64(scanDetail.TotalVersions)
-		sd.ProcessedKeys += int64(scanDetail.ProcessedVersions)
-		sd.RocksdbDeleteSkippedCount += scanDetail.RocksdbDeleteSkippedCount
-		sd.RocksdbKeySkippedCount += scanDetail.RocksdbKeySkippedCount
-		sd.RocksdbBlockCacheHitCount += scanDetail.RocksdbBlockCacheHitCount
-		sd.RocksdbBlockReadCount += scanDetail.RocksdbBlockReadCount
-		sd.RocksdbBlockReadByte += scanDetail.RocksdbBlockReadByte
-	}
 }
 
 const (
@@ -435,7 +317,7 @@ type CopRuntimeStats struct {
 	// same tikv-server instance. We have to use a list to maintain all tasks
 	// executed on each instance.
 	stats      map[string][]*basicCopRuntimeStats
-	scanDetail *ScanDetail
+	scanDetail *util.ScanDetail
 	// do not use kv.StoreType because it will meet cycle import error
 	storeType string
 }
@@ -739,7 +621,7 @@ func (e *RuntimeStatsColl) GetOrCreateCopStats(planID int, storeType string) *Co
 	if !ok {
 		copStats = &CopRuntimeStats{
 			stats:      make(map[string][]*basicCopRuntimeStats),
-			scanDetail: &ScanDetail{},
+			scanDetail: &util.ScanDetail{},
 			storeType:  storeType,
 		}
 		e.copStats[planID] = copStats
@@ -769,7 +651,7 @@ func (e *RuntimeStatsColl) RecordOneCopTask(planID int, storeType string, addres
 }
 
 // RecordScanDetail records a specific cop tasks's cop detail.
-func (e *RuntimeStatsColl) RecordScanDetail(planID int, storeType string, detail *ScanDetail) {
+func (e *RuntimeStatsColl) RecordScanDetail(planID int, storeType string, detail *util.ScanDetail) {
 	copStats := e.GetOrCreateCopStats(planID, storeType)
 	copStats.scanDetail.Merge(detail)
 }
